split up Expertise into different components
